Use Type System Types in Graph Service (#815)
* Fix Type System entry-points for tsc (#813)

* Fix exports in type-system package

* Add changeset

* Version Packages (#814)


* Move entity definitions into their own file

* Move file definitions into their own file

* Move link definitions into their own file

* Move linked aggregation definitions into their own file

* Move entity-type definitions into their own file

* Move block graph definitions into their own file

* Use Type System package for entity type definitions

* Disable entity type mutation methods

* Introduce `EntityId` type alias

* Add `EntityPropertiesObject` and `EntityMetadata`
