+++ packages/@blockprotocol/graph/package.json
@@ -57,6 +57,7 @@
   },
   "dependencies": {
     "@blockprotocol/core": "0.0.12",
+    "@blockprotocol/type-system": "0.0.2",
     "lit": "^2.4.1"
   },
   "devDependencies": {
