+++ packages/@blockprotocol/graph/src/types/file.ts
@@ -1,0 +1,15 @@
+import { EntityId } from "./entity";
+
+export type FileMediaType = "image" | "video";
+
+export type UploadFileData = {
+  file?: File | null;
+  url?: string | null;
+  mediaType: FileMediaType;
+};
+
+export type UploadFileReturn = {
+  entityId: EntityId;
+  url: string;
+  mediaType: FileMediaType;
+};
