+++ packages/@blockprotocol/graph/src/types/linked-aggregation.ts
@@ -1,0 +1,33 @@
+import {
+  AggregateEntitiesResult,
+  AggregateOperationInput,
+  Entity,
+} from "./entity";
+
+export type LinkedAggregationDefinition = {
+  aggregationId: string;
+  sourceEntityId: string;
+  path: string;
+  operation: AggregateOperationInput;
+};
+
+export type LinkedAggregation = Omit<LinkedAggregationDefinition, "operation"> &
+  AggregateEntitiesResult<Entity>;
+
+export type GetLinkedAggregationData = {
+  aggregationId: string;
+};
+
+export type CreateLinkedAggregationData = Omit<
+  LinkedAggregationDefinition,
+  "aggregationId"
+>;
+
+export type UpdateLinkedAggregationData = {
+  aggregationId: string;
+  operation: LinkedAggregationDefinition["operation"];
+};
+
+export type DeleteLinkedAggregationData = {
+  aggregationId: string;
+};
