+++ packages/@blockprotocol/hook/package.json
@@ -52,7 +52,8 @@
     "lint:tsc": "tsc --noEmit"
   },
   "dependencies": {
-    "@blockprotocol/core": "0.0.12"
+    "@blockprotocol/core": "0.0.12",
+    "@blockprotocol/graph": "0.0.18"
   },
   "devDependencies": {
     "@local/eslint-config": "0.0.0-private",
