+++ packages/@blockprotocol/type-system/CHANGELOG.md
@@ -1,0 +1,7 @@
+# @blockprotocol/type-system
+
+## 0.0.2
+
+### Patch Changes
+
+- [#813](https://github.com/blockprotocol/blockprotocol/pull/813) [`22991ad`](https://github.com/blockprotocol/blockprotocol/commit/22991ad7c63ad8683ce09f5a10743e481d6b474a) Thanks [@Alfred-Mountfield](https://github.com/Alfred-Mountfield)! - Fix the typescript exports of the package
