+++ packages/@blockprotocol/type-system/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@blockprotocol/type-system",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Definitions of types within the Block Protocol Type System",
   "homepage": "https://blockprotocol.org",
   "repository": {
@@ -34,6 +34,13 @@
   "main": "./dist/umd/index.js",
   "module": "./dist/es/index.js",
   "types": "./dist/umd/index.d.ts",
+  "typesVersions": {
+    "*": {
+      "slim": [
+        "./dist/cjs-slim/index-slim.d.ts"
+      ]
+    }
+  },
   "files": [
     "dist"
   ],
